Tempest: Intelligent image detection/creation
When an instance is deployed using the SSP disk driver with a glance
image named 'foo', we optimize by making the instance's boot disk a
linked clone to an image LU named 'image_foo_{md5sum}', where {md5sum}
is the md5 hash of the image file itself.  We only upload the image LU
if it doesn't already exist.

For a tempest run, we would like to ensure that the image LU already
exists in the SSP so that each instance deploy can simply link to it
rather than having to re-upload for every test.

This change set adds logic to os_ci_tempest.sh to determine whether the
right image LU already exists in the SSP.  If it does, it was named for
a certain image at some point, which may have been a different devstack,
on a different slave node, etc.  We glean the name of that image by
parsing the name of the LU, and then use that name for the glance image
we detect or create via prep_glance_image.

This change set also adds validation of the glance image's checksum.

If an appropriate image LU is not found, we would ideally like
os_ci_tempest.sh to create it.  However, due to bugs in the current
revision of pvmctl, this is not feasible.  Therefore, if we don't find
an appropriate image LU, we instead print a helpful message telling the
user how to create it manually.

Change-Id: I0032f95e0851ae6dda35e071d4a5f55700239b09