os_ci_tempest.sh: Correct IMGNAME and error exit
Correct two bugs in os_ci_tempest.sh:

1) The prep_for_tempest method intended to use the global $IMGNAME
variable (the path of the VM image file to upload) but, due to a
previous refactor, was using $imgname instead.

2) Error check failures go through a function called 'bail', which is
supposed to exit the whole script via 'exit 255'.  However, some of the
functions which did their error checking this way were being executed in
backticks - i.e. in a subshell - so the 'exit' was only causing
termination of that subshell, not the overall script.

Change-Id: I8c45731fdb1c73cee21e2d0c84db9103a87e8dd4