Uninterpolated single quotes, proper outfile path
1) The 'realpath' command used to work regardless of whether the target
existed.  Now it doesn't.  So attempting to resolve the output file's
path requires a little extra finagling.

2) MVCMD="mv -f '$TEMPEST_CONF_GEN' '$TEMPEST_CONF_INST'"
This should interpolate the variables upon definition.  Then when
executed as
$MVCMD
it should look like
mv -f '/path/to/whatever' '/path/to/otherthing'
Which, when interpolated by the shell, should use those single quotes to
isolate the path vars, in case they have spaces.  Instead, apparently,
it treats them as single quotes, so mv is literally trying to resolve
paths with single quotes in them.

Change-Id: I1a3212d56bc974455fda810637c33ccfe343120d