Create LPAR to prime the SSP with the right image
Previously, if the VM image didn't already exist in the SSP,
os_ci_tempest.sh would bail and the consumer would have to create it
manually.

With this change set, os_ci_tempest.sh will create an LPAR with the
appropriate image and wait until its LU has begun to upload.  Due to the
SSP driver logic in nova-powervm, other VMs can be created while this is
happening - they'll wait until the first upload is finished and reuse
the same image.  Thus the first VM-creating tests may take a while to
start, but not longer than they otherwise would have if they had to
upload the image themselves.  And subsequent tests reuse the image, so
run faster.

Change-Id: I2f0f3a9aeb8b09704bb6c481cce9b7e0200243e5