Specify net for primer LPAR, check nova boot RC
In a recent deploy, the nova boot command failed with this error:

17:10:17 ERROR (Conflict): Multiple possible networks found, use a
Network ID to be more specific. (HTTP 409) (Request-ID:
req-a9736330-d0df-4fdd-bbf0-b5c735312b64)

Then the following repeated "forever":
17:10:17 Looking for an existing Image LU with checksum
'b3348159ae7c7a7e949794a1f4194525'
17:10:31 No matching Image LU found.

With this change set, we a) check the exit code from the nova boot
command, and bail if it fails; and b) explicitly specify the network on
which the primer VM should be created.

Change-Id: I967e0e80d2fc45133d26efe6b125e78a0ecbbfd5