Turn off the shelve/unshelve test
The shelve/unshelve test completes properly on PowerVM.  But it runs
best when concurrency is set to 1.  The CI needs to run with concurrency
set to 4 to complete in time.  This may cause each test to run a little
longer, but you're running four at once.  Overall it cuts our time way
down (about in half).

However, since each test case runs a little slower, we are finding that
the shelve/unshelve test is reliable about 50% of the time.  It is
usually just outside the window of completion.

For now, we need to turn it off and revisit why it was taking so long in
a subsequent update.

Change-Id: Iae434ab4a31091e24bf6e0760ccda7fc15d457c6