Convert skip list from name to id
This converts the list of tests to skip from the test name to the
idempotent id of the test. If the name of the test gets changed, the
id will remain the same, meaning that test will still be found in the
skip list. It also cleans up the file, removing some non-existent tests,
tests that are duplicated, etc.

Change-Id: Iae49b51895b5787a79d0d98ff2f0191a024fe886