Disable incompatible tempest test
This test is failing for stable/newton runs. The problem is that
that we are checking out tempest tag 13.0.0 for newton runs while
the ceilometer tempest test assumes we are using a newer version
of tempest. This leads to incompatible methods. This is being disabled
until it can be debugged in the staging environment.

Change-Id: Ic9d400557cc236b6a2f24b9ca526eb25c4483b0c