Disable test with duplicate id
There are two tests with the same id, one in the tempest tests, one
in the neutron tempest tests. The neutron version of the test is
currently disabled and being investigated due to failures. But since
the tempest version of the test shares the id, it is still being run.
This adds the tempest version to the disabled tests as well.

Change-Id: I18adf1a4c504b0d2194c24118dc505a0e020a9b3