Add devstack configuration directory
This adds a devstack configuration directory to powervm-ci that will
house the local.conf files used for PowerVM CI. These files were
previously found in the neo-os-ci repo and passed into the CI env.
using ansible playbooks. The problem with that method was that changes
to the local.conf files could not be picked up by the CI instantly.
The ready nodes would have to be cleared and a new image template
would have to be built in order for the changes to propogate through
to the CI runs. By moving them to this repo, CI will be able to pull
the latest local.conf version every run.

Change-Id: I737ba417c44c71f1e77e4f1e48205a707a4b3376