Patch any devstack project
This sets up the infrastructure to patch any openstack project by
specifying pairs of repos and change number lists. This will allow
much greater flexibility in the CI environment. When patches are added
to a patching.conf file they will be applied the next CI run. We will
now be able to apply workarounds and test patches with more control and
less disruption.

Change-Id: I9286735fd21a4b49f006c4a5f7d32edb9f6fb573