Delete private network after in-tree stack
It was discovered when removing neutron legacy that devstack was
creating a private network on the master out-of-tree stack even though
we have the neutron setting for creating networks set to false. This
is also true of the master in-tree stack. This network is now being
deleted for both master stacks.

Change-Id: Ib40eadde675689a1d053f72d3e22582265c03fd2