Add timeout check to discover_hosts
The previous logic would run discover hosts until either a host was
found or it had reached a timeout of 60 seconds without finding any.
If no host was found, the script would not fail. This adds a timeout
check and causes the script to fail, killing the jenkins job, in the
event that no host is found.

Change-Id: I3df4292eff96f27f6e34978de5578e5efe5a021e