Reverse the order of project patching logic
This moves the openstack patching logic to a point later in the script
after the patch being tested is checked out. Since the tested patch is
a checkout and the openstack patches are cherry-picks, it must occur
in this order. Otherwise the checkout of the patch being tested
overwrites the patches that were already cherry-picked.

Change-Id: I2e970d9d3773e539c9236332573f4b261eaa3b74