Set network api extensions
This generates and sets api_extensions for networking to the extensions
that are enabled after stacking. This will allow us to start removing
tests from the skip lists and begin relying on the tempest logic for
determining which tests should be skipped. Eventually this entire file
will be eliminated in favor of a devstack generated tempest.conf. At
that point devstack will maintain the list of api extensions for us
based on our stack. Follow on changes for the other services api
extensions will be introduced after they are tested.

Change-Id: I9d3b4690fea35b98c76e8836f0b6b610d4b7ef78