Remove router tests from skip list
A recent change was introduced to our tempest configuration to specify
which networking api extensions are enabled. With this change, it is
no longer necessary to keep router tests in the skip list. We can now
rely on the tempest logic for determining which tests to skip because
they are either testing or they require routers. However there are a
fair number of tests that use routers without doing that check. These
tests will remain in the skip list until the bugs can be identified
and fixed.

Change-Id: I02b452dbd9508d4ec46eed8ab055b22fdcc55d67