Update local.conf files for devstack generated tempest.conf
This is the first step in creating devstack generated tempest
configurations. It uses the existing local.conf settings for generating
tempest configurations. There are further changes that will need to be
made to the local.conf files, but they require changes to devstack
and/or tempest. Once that external work is done, the remaining changes
to the local.conf files can be made and the CI logic can be rewritten
to use the new generated tempest configuration.

Change-Id: Id1e6bb99b5f96881a52c0d0030c05f6ddb4bae61