Increase discover_hosts timeout
The current timeout waiting for the discover_hosts command to succeed
is 60 seconds. Typically the host is discovered within the first 10
seconds. However some CI runs are failing to discover the host within
the 60 second window. When testing manually I found an instance that
did not discover a host within 1 min. but was able to find one within
2 min. This increases the timeout to 3 min to add a buffer on top of
those findings. This value may need to be further adjusted depending
on the results at volume.

Change-Id: I1d80988979355af1e63bede9fcb9c205308822d7