Set EnvironmentFile for services pre-stack
This sets the EnvironmentFile variable for the services that require
remote pypowervm prior to stacking. This environment file contains a
path to the session config needed for remote pypowervm to work.
Previously this was getting set after stacking at which point the
services would be restarted for the change to take affect. By adding
this setting before stacking the services can start up properly during
the stack. This also eliminates the need to run discover_hosts after
stacking because the hosts get discover properly while stacking.

Change-Id: I08252d9f000dfc593b09ffeb40b77c7ad526c0d8