Clone pypowervm if patching
Recently the image template build script was changed to install
pypowervm from pypi rather than cloning and installing the git repo.
This means that /opt/stack/pypowervm doesn't exist when prep_devstack
is run. This breaks any runs that use the pypowervm patching logic.
This change will clone pypowervm to /opt/stack/pypowervm for any runs
that need to apply patches to it.

Change-Id: I062ee8b14d6db95a68dd8707cb7eef7209ed209b