Discover hosts after stacking
The stack.sh script runs discover_hosts which usually finds the hosts
correctly. However it does occasionally fail to find any hosts while
stacking. This is due to a race condition between the compute service
starting and discover_hosts being run that should be fixed by [1]. This
adds a second discover_hosts after the stack that should find any hosts
that weren't discovered initially. We can leave this in after [1] is
merged as a fallback in case the first discover_hosts fails to find the
hosts for a different reason.

[1] https://review.openstack.org/#/c/488381/

Change-Id: Ifc1adbd7b205c37e1fd25b052b134caf06851433