Apply retry show_server on timeout tempest patch
This applies [1] to tempest for each master run. The REST request in
show_server is intermittently timing out on specific tests and causing
them to fail. The patch will retry the request until it either succeeds
or the test times out. This will allow us to determine if the issue is
temporary or not.

Change-Id: I2afdc6e6f039fc8885a2b02816c26d0b16a5597c