Move process_logs to powervm-ci
The process_logs.sh script was previously found in the internal
neo-os-ci repo. By moving the script to powervm-ci, any further changes
to the script will be picked up instantly by the next CI run. Stopping
all devstack services and outputting the journalctl logs to files are
also included in the process_logs script instead of being defined in the
jenkins job configuration

Change-Id: I9f6ba00cf27bd4f29ca43e8f59bd0d1011fc0141