Process logs cleanup
This updates the zip to be done on .txt files instead of .log files.
It also changes the extension on the tempest and stack logs so that
they are picked up by the zip and can be displayed in browser. Lastly
it adds the apache2 logs to those copied to the logserver.

Change-Id: If1e385fd6dfd82080ee0183d52a591cf6f41375f