Remove pvm-q-sea-agt from in-tree enabled services
pvm-q-sea-agt was listed as one of the enabled services in the in-tree
local.confs. This isn't actually being used and is being removed from
the in-tree local.confs. It is causing the in-tree stacks to fail since
the full removal of screen support [1], which introduced a check to
confirm that all of the services listed as enabled are enabled and
running.

[1] https://review.openstack.org/#/c/499186/

Change-Id: I3673fc99618c68c7d3d4f94a1b79f016fa2d2c03