Revert "Revert "Discover hosts after stacking""
We thought that the merge of [1] would remove the need to discover_hosts
after stacking. However we are still seeing runs that fail to discover
the compute host during the stack, causing the associated tempest run to
fail. This restores the post-stack discover_hosts command while we
determine what the issue is.

[1] https://review.openstack.org/#/c/488381/

This reverts commit ff34ef61d931df52030769a962474f80dcf7a216.

Change-Id: I46db1fedb849a22b29c2843dd7aca4d9a291b554