Merge patch being tested
The logic prior to this change would checkout the patch that CI was
running against and use that for the CI run. It should actually be
merging the patch with the latest code for the branch being run
against. This will not cause any merge conflicts because zuul-merger
checks for conflicts prior to the prep_devstack script being run.

Change-Id: I4ace18870fa4870b3590f71dbb6c68268aa1162e