Replace in-tree instance_name_template
The prep_devstack.sh script replaces the instance_name_template value
in the local.conf files before stacking. This has been occuring both
in and out-of-tree. However, the in-tree local.conf files had a usable
template value in the local.conf files. This replaces those values
with REPLACE_THIS_BEFORE_STACKING to clarify that the templates are
being generated each run. The template value includes the vm id which
is discovered each run in prep_devstack.

Change-Id: I77aeb50e5d6e58f03f3e6b15bde36b49db9945df