Remove journalctl log coloring
This change removes the log color codes when outputting the devstack
service journals to files. The logserver currently expects the txt.gz
files to be free of color codes when converting to the html format. By
removing the color codes the logs will now be highlighted based on
severity when viewed in a browser. However this will remove the ability
to view colorized logs if downloading and viewing the source files. A
solution allowing colorized logs for both use cases will be explored
in the future.

Change-Id: I5c6634132129669e8c5c3d878170289149cce28a