Update openstack repo setup
This updates the setup of the openstack repos. First, it replaces the
git fetch and merge of the change being tested with a git pull. Second
it updates the patching logic to properly determine if the changes in
the patching.conf files are already in the commit chain. If so it will
skip applying those patches.

Change-Id: I6ca326b366243efe72f1910e9ac5470638468cdf