Swizzle instance name template to avoid conflicts
In Tempest, instance names can frequently have the same first 11
characters.  With the old instance name template, this left only the
first 8 hex digits of the UUID to guarantee uniqueness of things like
VOpt names -- which may not have been enough.

With this change, we front-load the instance name template with the full
instance UUID - which means we ought to get most or all of it in the
VOpt name - and put the display name at the back.  This should reduce
or eliminate the probability of VOpt name collisions, which may resolve
a weird bug we were seeing where a scrubber would occasionally try to
remove the VOpt being created by a spawn.

Change-Id: I69a08da92efb78748fc255058cd0c3272dce23e9