Patch name vopt by uuid
This will apply [1] to all master out-of-tree CI runs. We suspect that
by naming the vopts using the UUID, we should avoid naming collisions
causing tempest failures. The collisions are pretty infrequent. This
will temporarily be applied to all master OOT runs to get a high
enough volume of runs to confirm that the patch alleviates the issue.

[1] https://review.openstack.org/#/c/540453/

Change-Id: I0ff9450f8573fbd68ecf084a71a80992481b865c