Patch experimental proc_units_factor hardcode for IT runs
In-tree runs are sometimes failing to create lpars because there aren't
enough available processing units. It is suspected that this is due to
different defaults for the proc_units_factor between in-tree and
out-of-tree runs. Out-of-tree uses the nova-powervm default of 0.1
and in-tree uses the pypowervm default of 0.5. [1] hardcodes the value
to 0.1 in-tree. By applying this patch to all IT master runs we should
be able to determine if this actually the issue.

[1] https://review.openstack.org/#/c/554688/

Change-Id: Iaf1d84e64e66e4bc1a40e8ced6a95a54f101650b