Enable tempest snapshot testing against in-tree snapshot change
This is a temporary change that will allow our CI to run snapshot
testing on the in-tree tempest snapshot change [1] until it is merged.
It checks for the snapshot change number each run and removes the line
that disables snapshot testing from the local.conf if it matches. This
allows us to get snapshot testing in our CI without patching every CI
run with the in-tree snapshot change.

[1] https://review.openstack.org/#/c/543023/
Change-Id: Ib4d0a24c71aaa898b3a3fbb8a8ea83c64e8915bf