Scenario testing
This enables two tempest scenario tests for PowerVM CI. Enabling
scenario testing is a requirement for the nova in-tree driver work in
the Rocky development cycle. The neutron networks and tempest settings
have been updated so that it is possible for the tempest vms to reach a
pingable state. Run validation has only been enabled for master
branches. Similar changes will be required for the stable branches.

Change-Id: Ife700947e2f795f4fa6b1f0323de4e5dd2afc331