Only copy the jenkins console log if build_url is provided
The process_logs script copies the jenkins console log to the
logserver. Since jenkins is not involved in manual CI runs, the jenkins
console log does not exist. This allows the user to leave out the
final parameter (build_url) when executing the script to avoid trying
to process a non-existent log file.

Change-Id: I03a432af82a34484ba1f98e3e42a51dbd7b1b2a2