vSCSI FC CI
This is the starting point for vSCSI Fibre Channel CI. It introduces
a vSCSI specific local.conf, patching.conf, and tempest configuration.
vSCSI CI will run a set of three scenario tests to ensure that volumes
can be attached and detached to the tempest vms. It only covers master
OOT changes, further work will be done to add vSCSI CI for IT and stable
branches.

Change-Id: If1b96f7e554dad349f88092b82669f9dc84ed749