Enables scenario testing on stable branches
Enables scenario testing for stable branches (in-tree and out-of-tree)
to confirm that tempest VMs are able to reach a pingable state. This
satisfies a nova core requirement. See also commit
f0d314e0e13b7b51591035fbae3f1b87ecf5967f which added scenario tests for
the master branches. This does not address in-tree stable/pike because
that release did not support networking for VMs. This includes a patch
to allow password authentication for the SSH test until we are able to
figure out why keypair auth is not working.

Change-Id: I9f2a72cc0395397bef34ee20374f11c29e2cfe25